+++ app/src/main/java/com/survivalcoding/ifkakao/data/datasource/ConferenceService.kt
@@ -1,6 +1,6 @@
 package com.survivalcoding.ifkakao.data.datasource
 
-import com.survivalcoding.ifkakao.domain.model.ApiDto
+import com.survivalcoding.ifkakao.data.dto.ApiDto
 import retrofit2.Response
 import retrofit2.http.GET
